Clean up provider APIs
